pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "fastapi-mongo-base"
version = "1.0.45"
description = "A simple boilerplate application, including models and schemas and abstract router, for FastAPI with MongoDB"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
license-files = ["LICENSE.txt"]
keywords = ["fastapi", "mongodb", "beanie"]
authors = [{ name = "Mahdi Kiani", email = "mahdikiany@gmail.com" }]
maintainers = [{ name = "Mahdi Kiani", email = "mahdikiany@gmail.com" }]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Build Tools",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3 :: Only",
]
dependencies = [
  "pydantic>=2.12.3",
  "httpx>=0.28.1",
  "singleton_package>=0.8.2",
  "json-advanced>=0.12.11",
  "pytz>=2025.2",
  "beanie>=1.30.0",
  "fastapi>=0.119.0",
  "uvicorn[standard]>=0.38.0",
  "uuid6>=2025.0.1",
]

[project.optional-dependencies]
usso = ["usso>=0.28.47"]
sql = ["sqlalchemy>=2.0.43"]
test = ["pytest", "pytest-cov", "pytest-asyncio", "mongomock_motor", "coverage"]

[project.urls]
"Homepage" = "https://github.com/mahdikiani/fastapi-mongo-base-app"
"Bug Reports" = "https://github.com/mahdikiani/fastapi-mongo-base-app/issues"
"Funding" = "https://github.com/mahdikiani/fastapi-mongo-base-app"
"Say Thanks!" = "https://saythanks.io/to/mahdikiani"
"Source" = "https://github.com/mahdikiani/fastapi-mongo-base-app"

[tool.setuptools]
package-data = { "usso" = ["*.dat"] }

[tool.ruff]
line-length = 79
target-version = "py313"
fix = true
unsafe-fixes = true
preview = true

[tool.ruff.lint]
select = [
  "F",     # Pyflakes: critical errors and unused code
  "E",
  "W",     # PEP8: formatting issues
  "I",     # isort: import sorting
  "UP",    # pyupgrade: modern syntax
  "B",     # bugbear: potential bugs
  "T",     # debugger / print statement checks
  "C",     # complexity: complexity checking
  "A",     # avoids shadowing Python built-ins
  "ASYNC", # async/await syntax
  "SIM",   # simplify: cleaner code
  "PERF",  # performance optimizations
  "TRY",   # better try/excepts
  "N",     # naming conventions
  "S",     # security checks
  "ANN",   # type annotation checks
  "T20",   # no print()
  "C4",    # comprehension: better loops
  "COM",   # enforces trailing commas where needed.
  "EXE",   # executable: executable checking
  "ISC",   # implicit string concatenation
  "G",     # misuse of logging string interpolation
  "D",     # docstrings: docstring checking
  "ERA", # remove commented code
  "RUF", # ruff-specific improvements
]
per-file-ignores = { "tests/*" = ["G004", "S101"], "server/db.py" = ["TRY"] }
ignore = ["TRY003", "COM812", "B008", "D202", "D203", "D209", "D212", "D413"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true
check_untyped_defs = false
disallow_untyped_defs = false
